Reworked Implementation structure to more easily swap implementations of solutions, is not final mechanism

+++ src/main/java/ch/grosscy/exercises/challenge05/Solution05.java
@@ -2,7 +2,7 @@ package ch.grosscy.exercises.challenge05;
 
 public class Solution05 implements Challenge05 {
     @Override
-    public long calculate(int upperLimit) {
+    public long calculate(long upperLimit) {
         // TODO Implementation...
         return 0L;
     }
