+++ src/main/java/ch/grosscy/exercises/challenge06/Solution06.java
@@ -2,7 +2,7 @@ package ch.grosscy.exercises.challenge06;
 
 public class Solution06 implements Challenge06 {
     @Override
-    public long calculate(int upperLimit) {
+    public long calculate(long upperLimit) {
         // TODO Implementation...
         return 0L;
     }
