+++ src/main/java/ch/grosscy/exercises/challenge10/Solution10.java
@@ -2,7 +2,7 @@ package ch.grosscy.exercises.challenge10;
 
 public class Solution10 implements Challenge10 {
     @Override
-    public long calculate(int upperLimit) {
+    public long calculate(long upperLimit) {
         // TODO Implementation...
         return 0L;
     }
