+++ src/main/java/ch/grosscy/exercises/challenge11/Solution11.java
@@ -3,7 +3,7 @@ package ch.grosscy.exercises.challenge11;
 public class Solution11 implements Challenge11 {
 
     @Override
-    public long calculate(int searchRayLength) {
+    public long calculate(long searchRayLength) {
         // TODO Implementation...
         return 0L;
     }
