+++ src/main/java/ch/grosscy/exercises/challenge14/Solution14.java
@@ -3,7 +3,7 @@ package ch.grosscy.exercises.challenge14;
 public class Solution14 implements Challenge14 {
 
     @Override
-    public long calculate(int upperLimit) {
+    public long calculate(long upperLimit) {
         // TODO Implementation...
         return 0L;
     }
