+++ src/main/java/ch/grosscy/exercises/challenge15/Solution15.java
@@ -4,7 +4,7 @@ package ch.grosscy.exercises.challenge15;
 public class Solution15 implements Challenge15 {
 
     @Override
-    public long calculate(int gridSize) {
+    public long calculate(long gridSize) {
         // TODO Implementation...
         return 0L;
     }
