+++ src/main/java/ch/grosscy/exercises/challenge16/Solution16.java
@@ -4,7 +4,7 @@ package ch.grosscy.exercises.challenge16;
 public class Solution16 implements Challenge16 {
 
     @Override
-    public long calculate(int exponent) {
+    public long calculate(long exponent) {
         // TODO Implementation...
         return 0L;
     }
